+++ Configuration/TSconfig.ts
@@ -1,4 +1,0 @@
-# === Includes ===
-
-<INCLUDE_TYPOSCRIPT: source="FILE: ./TSconfig/GridElements.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./TSconfig/ContentElements.ts">