+++ Configuration/TSconfig/GridElements.ts
@@ -1,8 +1,0 @@
-# === Includes ===
-
-<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/AccordionGridElement.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/ColumnLayoutGridElement.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/DocumentSectionGridElement.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/PanelGridElement.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/TimelineGridElement.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./GridElements/ViewportGridElement.ts">