+++ Configuration/TSconfig/GridElements/ColumnLayoutGridElement.ts
@@ -1,4 +1,0 @@
-# === Includes ===
-
-<INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/2Columns.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/3Columns.ts">