+++ Configuration/TypoScript/GridElements/ColumnLayoutGridElement/constants.ts
@@ -1,4 +1,0 @@
-# === Includes ===
-
-<INCLUDE_TYPOSCRIPT: source="FILE: ./2Columns/constants.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./3Columns/constants.ts">