+++ Configuration/TypoScript/GridElements/setup.ts
@@ -1,8 +1,0 @@
-# === Includes ===
-
-<INCLUDE_TYPOSCRIPT: source="FILE: ./AccordionGridElement/setup.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./ColumnLayoutGridElement/setup.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./DocumentSectionGridElement/setup.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./PanelGridElement/setup.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./TimelineGridElement/setup.ts">
-<INCLUDE_TYPOSCRIPT: source="FILE: ./ViewportGridElement/setup.ts">